Enables usage of multiple audio samples used for voice cloning.
Previously, users were able to use a single audio sample to synthesize voice. This update enables users to specify a directory with multiple audio samples that will all be used toward voice cloning.

In the Coqui docs, this is accomplished by using a regex, but I thought this overly complicated the typical use case--provide a single audio file, or a directory with multiple audio files.